feature: Save usernames retrieved by Twitch
We stopped to retrieve the usernames via the heat extensions because of a lot of failures.

And to avoid calling the Twitch API too much, we store the user the first time we encounter a new ID.

In addition to that we also have a new "allowed" field on the Player model to disallow baned users to play.